+++ circle-test.sh
@@ -54,10 +54,10 @@ case $CIRCLE_NODE_INDEX in
         go test $PARALLELISM $TIMEOUT -v ./... 2>&1 | tee $CIRCLE_ARTIFACTS/test_logs.txt
         rc=${PIPESTATUS[0]}
         ;;
-    1)
-        GORACE="halt_on_error=1" go test $PARALLELISM $TIMEOUT -v -race ./... 2>&1 | tee $CIRCLE_ARTIFACTS/test_logs_race.txt
-        rc=${PIPESTATUS[0]}
-        ;;
+    #1)
+    #    GORACE="halt_on_error=1" go test $PARALLELISM $TIMEOUT -v -race ./... 2>&1 | tee $CIRCLE_ARTIFACTS/test_logs_race.txt
+    #    rc=${PIPESTATUS[0]}
+    #    ;;
 esac
 
 exit $rc